refactor(admin): extraer lógica de cierre de sesión a hook reutilizable
