+++ src/components/admin/layout/AdminHeader.jsx
@@ -1,27 +1,8 @@
-import { useNavigate } from "react-router-dom";
-import { clearAuth } from "../../../utils/authStorage";
 import "./AdminHeader.css";
-import Swal from "sweetalert2";
+import useLogout from "../hooks/useLogout";
 
 export default function AdminHeader({ Icon, ICONS }) {
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    Swal.fire({
-      title: "¿Cerrar sesión?",
-      text: "Tu sesión actual finalizará.",
-      icon: "question",
-      showCancelButton: true,
-      confirmButtonText: "Cerrar sesión",
-      cancelButtonText: "Cancelar",
-      confirmButtonColor: "#111827",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        clearAuth();
-        navigate("/Login", { replace: true });
-      }
-    });
-  };
+  const { handleLogout } = useLogout();
 
   return (
     <header className="admin-header">
